+++ src/versioning/Versioning.ts
@@ -13,6 +13,6 @@ export class Versioning {
    * Return string representation of SDK version
    */
   static get sdkVersion(): string {
-    return '2.11.1';
+    return '2.12.0';
   }
 }